requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.0
django-cleanup==6.0.0
django-jet==1.0.8
django-jet-reboot==1.3.1
docopt==0.6.2
feedparser==6.0.10
idna==3.4
Pillow==9.4.0
python-decouple==3.7
requests==2.28.2
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
yarg==0.1.9
